Moved all api logic to the controlllers directory,
added a authenticated delete user route

fixed no css for the components on privacy and terms
