+++ public/terms/index.html
@@ -10,6 +10,9 @@
     
     <title>Terms of Service</title>
     <link rel="stylesheet" href="/css/index.css">
+    <link rel="stylesheet" href="/components/header/topbar.css">
+    <link rel="stylesheet" href="/components/footer/footer.css">
+    <link rel="stylesheet" href="/components/settings/settings-modal.css">
     <script src="/js/theme-init.js"></script>
     <!-- Shared components -->
     <script src="/components/header/topbar.js"></script>
@@ -19,7 +22,7 @@
     <script src="/components/cookies/cookie-banner.js"></script>
 </head>
 <body>
-    <topbar-component auth-state="guest" active-page="home"></topbar-component>
+    <topbar-component auth-state="guest" active-page="terms"></topbar-component>
     <script>
         document.addEventListener('DOMContentLoaded', () => {
             const topbar = document.querySelector('topbar-component');
